core dump updates.  change vfiles to use 32 bit lengths

+++ src/lib_battery/battery.c
@@ -118,9 +118,9 @@ static uint8_t fx_low_batt[] __attribute__((aligned(4))) = {
     #include "low_batt.fx.carray"
 };
 
-static uint16_t fx_low_batt_vfile_handler( vfile_op_t8 op, uint32_t pos, void *ptr, uint16_t len ){
+static uint32_t fx_low_batt_vfile_handler( vfile_op_t8 op, uint32_t pos, void *ptr, uint32_t len ){
 
-    uint16_t ret_val = len;
+    uint32_t ret_val = len;
 
     // the pos and len values are already bounds checked by the FS driver
     switch( op ){
@@ -145,9 +145,9 @@ static uint8_t fx_crit_batt[] __attribute__((aligned(4))) = {
     #include "crit_batt.fx.carray"
 };
 
-static uint16_t fx_crit_batt_vfile_handler( vfile_op_t8 op, uint32_t pos, void *ptr, uint16_t len ){
+static uint32_t fx_crit_batt_vfile_handler( vfile_op_t8 op, uint32_t pos, void *ptr, uint32_t len ){
 
-    uint16_t ret_val = len;
+    uint32_t ret_val = len;
 
     // the pos and len values are already bounds checked by the FS driver
     switch( op ){
